Get MERMAID project sites

Get sites for specified MERMAID project(s). Returns metadata on sites, including site ID and name, notes, latitude and longitude, country, reef type and zone, exposure, and when the site was created and last updated. Requires authorization.

mermaid_get_project_sites(
  project = mermaid_get_default_project(),
  limit = NULL,
  token = mermaid_token()
)

Arguments

project
A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from mermaid_get_my_projects or mermaid_search_my_projects. Defaults to the projects listed via mermaid_get_default_project, if available.

limit
Number of records to get. Use NULL (the default) to get all records.

token
API token. Authenticate manually via mermaid_auth, or automatically when running any project- or user-specific functions (like this one).

Examples

if (FALSE) {
projects <- mermaid_get_my_projects()
projects %>%
  mermaid_get_project_sites()
}