Bulk validate records in Collecting

Bulk validates records in Collecting for a given project, and returns information on how many records produced errors, produced warnings, or were successfully validated without errors or warnings. To be used after mermaid_import_project_data.

mermaid_import_bulk_validate(project, token = mermaid_token())

Arguments

project

A way to identify the project(s). Can be project IDs (passed as a character vector directly) or projects resulting from mermaid_get_my_projects or mermaid_search_my_projects. Defaults to the projects listed via mermaid_get_default_project, if available.

token

API token. Authenticate manually via mermaid_auth, or automatically when running any project- or user-specific functions (like this one).

Examples

if (FALSE) { # \dontrun{
mermaid_get_my_projects() %>%
  head(1) %>%
  mermaid_import_bulk_validate()

# 43 records being validated...
# ✖ 27 records produced errors in validation
# • 13 records produced warnings in validation
# ✔ 3 records successfully validated without warnings or errors
} # }