Get or check field options for MERMAID import

Check the options available for a given field in MERMAID, either to see what the options are or to check that your data matches the available fields.

mermaid_import_field_options(
  field = c("width", "fishsizebin", "reefslope", "visibility", "current",
    "relativedepth", "tide", "fishname", "benthicattribute", "growthform",
    "habitatcomplexityscore"),
  values
)

Arguments

field

Field to check options for. One of: width, fishsizebin, reefslope, visibility, current, relativedepth, tide, fishname, benthicattribute, growthform, habitatcomplexityscore.

values

Optional values from your data, to match against valid options in Collect

Examples

if (FALSE) {
mermaid_import_field_options("growthform")
## A tibble: 11 x 1
#   mermaid_value
#   <chr>
# 1 Arborescent
# 2 Branching
# 3 Columnar
# 4 Corymbose
# 5 Digitate
# 6 Encrusting
# 7 Foliose
# 8 Massive
# 9 Mushroom coral
# 10 Plates or tables
# 11 Submassive

mermaid_import_field_options("growthform", c("Arboresent", "Branching", "Columar", "Foliose"))
## A tibble: 4 x 3
#   value      closest_mermaid_value match
#   <chr>      <chr>                 <lgl>
# 1 Arboresent Arborescent           FALSE
# 2 Branching  Branching             TRUE
# 3 Columar    Columnar              FALSE
# 4 Foliose    Foliose               TRUE
}