Search all MERMAID projects

Search all MERMAID projects by project name, country, or tags. Returns metadata on project, including ID, name, countries, number of sites, tags, notes, status, data sharing policies, and when the project was created and last updated.

mermaid_search_projects(
  name = NULL,
  countries = NULL,
  tags = NULL,
  include_test_projects = FALSE,
  limit = NULL
)

Arguments

name

Project name

countries

Project country. Projects are returned if the countries field contains countries, not just if it is exactly the same. For a list of countries used in MERMAID, see mermaid_countries

tags

Project tags. Projects are returned if the tags field contains tags, not just if it is exactly the same.

include_test_projects

Whether to include test projects. Defaults to FALSE.

limit

Number of records to get. Use NULL (the default) to get all records.

Examples

if (FALSE) { # \dontrun{
mermaid_search_projects(tags = "WCS Fiji")
mermaid_search_projects(countries = "Fiji", tags = "WWF-UK")

# The countries (or tags) do not have to be exactly the same
# A project is returned if it *contains* the countries/tags:
mermaid_search_projects(countries = "Tanzania", limit = 1)[["countries"]]

# To search within your projects only, use mermaid_search_my_projects():
mermaid_search_my_projects(countries = "Fiji")
} # }